Incorporate changes adopted in SEP 029
This commit updates the language of SEP 001 to be consistent with the proposed changes in SEP 029. Namely, it allows for SEP issues can be closed when action is no longer required. This commit also clarifies that the definitive text for an SEP is in the repository and not the issue tracker.